Correções de Bug do Perfil

+++ app/Http/Controllers/PerfilUserController.php
@@ -5,24 +5,25 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\{
-    Auth,Validator,Storage
+    Auth,Validator,Storage,Hash,
 };
 use App\Traits\PessoaTrait;
+
 
 
 class PerfilUserController extends Controller
 {
     use PessoaTrait;
+
+    //Metodo que retorna a views do Perfil
     public function index(){
         return view('perfil.perfil');
     }
 
+    //Metodo para Actualizar dados do Usuario
     public function update(Request $request)
     {   
-
-        // $test=Auth::user()->imagem_usuario;
-        // dd($test);
-        
+    
         $user= Auth::user();
         $user= User::findOrFail($user->usuario_id)->firstOrFail();
 
@@ -41,7 +42,7 @@ class PerfilUserController extends Controller
             'telefone_update'=>'required|size:9|unique:pessoas,telefone,'.$user->pessoa_id.',pessoa_id',
 
             //Formulario do user
-            'email_update' => 'required|email|max:200|unique:users,email,'.$user->usuario_id.',usuario_id',
+            'email_update'=> 'required|email|max:200|unique:users,email,'.$user->usuario_id.',usuario_id',
 
             //Formulario do endereço
             'municipio_update'=>'required|string|min:2',
@@ -49,6 +50,7 @@ class PerfilUserController extends Controller
             'zona_update'=>'required|string',
             'num_casa_update'=>'required|numeric',
         ];
+
         $msg_erro=[
 
             '*.required'=>'Este campo deve ser preenchido',
@@ -88,7 +90,6 @@ class PerfilUserController extends Controller
             'bairro_update'=>$request->bairro_update,
             'zona_update'=>$request->zona_update,
             'num_casa_update'=>$request->num_casa_update,
-
         ];
 
         $validator= Validator::make($dadosFiltrados,$regras_gerais,$msg_erro);
@@ -132,6 +133,7 @@ class PerfilUserController extends Controller
         return redirect()->route('perfil')->with('success', 'Dados actualizados com sucesso!');
     }
 
+    //Metodo para actualizar Imagem
     public function atualizarImagem($request)
     {   
         $user= Auth::user();
@@ -170,11 +172,60 @@ class PerfilUserController extends Controller
         }  
 
         if (file_exists($imagem_antiga)) {
+            if (!is_writable($imagem_antiga)) {
+                return redirect()->back()->with('erro_imagem_004', 'Lamentamos! Não foi possível excluir a imagem antiga.');
+            }
             unlink($imagem_antiga);
-            return redirect()->back()->with('success_imagem_001', 'Imagem atualizada com sucesso.');
         }
+        
+        return redirect()->back()->with('success_imagem_001', 'Imagem atualizada com sucesso.');
+        
     
     }
-  
+
+    //Metodo para Actualizar Senha
+    public static function changePassword(Request $request){
+
+        switch($request->password) {
+            case $request->password_confirmation:
+                goto conti_salvar;
+            default:
+            $msg = "A confirmação da senha não correspondem.";
+            return redirect()->back()->with('erro_senha_001', $msg)->withInput();
+        }
+
+        conti_salvar:
+        $dados=[
+            'password_old'=>$request->password_old,
+            'password'=>$request->password,
+        ];
+        $regras=[
+            'password' => 'required|min:6',
+            'password_old'=>'required|min:6',
+        ];
+        $msg_erro=[
+            '*.required'=>'Este campo deve ser preenchido',
+            'password.min'=>'A senha deve conter no minimo 6 letras',
+            'password_old.min'=>'A senha deve conter no minimo 6 letras',
+        ];
+        $validator= Validator::make($dados,$regras,$msg_erro);
+
+        if($validator->fails()){
+            return redirect()->back()->with('erro_senha_001', 'Erro senha não redifinida.')->withErrors($validator)->withInput();
+        }
+        $userAuth= Auth::user();
+        $user= User::findOrFail($userAuth->usuario_id);
+
+        if (!Hash::check($request->password_old, $user->password)) {
+            return redirect()->back()->with('erro_senha_002', 'Senha actual incorrecta.')->withInput();
+        }
+
+        $user->password = bcrypt($request->password);
+        if($user->save()){
+            return redirect()->back()->with('success_updatePassword_001','Senha redefinida com sucesso!');
+        }
+
+    }
+
 
 }
